Add files via upload
This is the second part of the program. This program asks the user for the file saved initially and then uses a recorder API to check the file content with the recorded content. Furthermore, the results of the program are posted in a tabular program with percentage of accuracy based on matching words.